+++ src/main/scala/org/mmisw/orr/ont/results.scala
@@ -50,7 +50,7 @@ case class OntologySummaryResult(
             orgName:        Option[String] = None,
             author:         Option[String] = None,
             status:         Option[String] = None,
-            metadata:       List[Map[String, AnyRef]] = List.empty,
+            metadata:       Option[List[Map[String, AnyRef]]] = None,
             ontologyType:   Option[String] = None,
             resourceType:   Option[String] = None,
             versions:       Option[List[String]] = None,